requirements.txt
blinker==1.9.0CacheControl==0.14.2cachetools==5.5.2certifi==2025.1.31cffi==1.17.1charset-normalizer==3.4.1click==8.1.8cmdstanpy==1.2.5colorama==0.4.6contourpy==1.3.0cryptography==44.0.2cycler==0.12.1firebase-admin==6.7.0Flask==2.2.5flask-cors==5.0.1fonttools==4.56.0google-api-core==2.24.2google-api-python-client==2.166.0google-auth==2.38.0google-auth-httplib2==0.2.0google-cloud-core==2.4.3google-cloud-firestore==2.20.1google-cloud-storage==3.1.0google-crc32c==1.7.1google-resumable-media==2.7.2googleapis-common-protos==1.69.2grpcio==1.71.0grpcio-status==1.71.0holidays==0.69httplib2==0.22.0idna==3.10importlib_metadata==8.6.1importlib_resources==6.5.2itsdangerous==2.2.0Jinja2==3.1.6joblib==1.4.2kiwisolver==1.4.7MarkupSafe==3.0.2matplotlib==3.9.4msgpack==1.1.0numpy==2.0.2packaging==24.2pandas==2.2.3pillow==11.1.0prophet==1.1.6proto-plus==1.26.1protobuf==5.29.4pyasn1==0.6.1pyasn1_modules==0.4.2pycparser==2.22PyJWT==2.10.1pyparsing==3.2.3python-dateutil==2.9.0.post0pytz==2025.2requests==2.32.3rsa==4.9scikit-learn==1.6.1scipy==1.13.1six==1.17.0stanio==0.5.1threadpoolctl==3.6.0tqdm==4.67.1tzdata==2025.2uritemplate==4.1.1urllib3==2.3.0Werkzeug==2.2.3zipp==3.21.0gunicorn  # Add this to avoid the 'gunicorn not found' error